+++ lib/src/platform_info_io.dart
@@ -1,5 +1,8 @@
+import 'dart:ffi';
 import 'dart:io';
 
 String get operatingSystem => Platform.operatingSystem;
 
 String get operatingSystemVersion => Platform.operatingSystemVersion;
+
+String get processArchitecture => Abi.current().toString();